refactor: extract app formatting helpers
